file_organizer/requirements.txt
pandas
python-docx
PyPDF2
python-pptx
openpyxl
streamlit
